Prevent Double Fetches on Admin Pages
When initializing the client for authorized customers visiting routes
that require authentication, wait for the ReAuthorize request to return
before fetching the page's data. This prevents double-fetching of the
page data for Admin routes as well as flashing of the "Access Denied"
page that appears before the ReAuthorize request resolves.