[#1525] Add Initial Structuring for AvaTax Integration
Add the basic scaffolding for our AvaTax API integration to a new
`Avalara` module.

The module currently supports client configuration(environment,
application name/version, & authorization), error handling, and the
simplest request type - the `ping` endpoint. The `req` library is used
to make the HTTP calls to AvaTax's API & helper types/functions are
included to ease the addition of new endpoints.

Tests are included for ensuring the proper decoding of the Error & Ping
responses.

Refs #1525