[#1525] Add CommitTransaction Request to the Avalara Integration
Add a `commitTransaction` request function to the Avalara module for
marking a Transaction as committed, allowing a saved Transaction to be
used for sales tax reporting calculations.

Fix a spelling error in the CreateTransactionRequest type name.

Add CompanyCode & TransactionCode types.

Use the CustomerCode type in the cCustomerCode field of the Customer
type.

Refs #1525