[#1525] Add Request for Voiding Avalara Transactions
Add a `voidTransaction` request to the Avalara integration module,
allowing us to mark a sales tax Transaction as void.

Since we need to create an uncommitted Transaction to obtain the final
order total for charging a Customer via Stripe, it's possible for a
route handler to fail after creating a Transaction but before committing
it(e.g., the Stripe request fails). In these situations, we can void the
uncommitted Transaction and it will be removed from the Avalara history.

Refs #1525