[#1525] Add TaxIncluded Field to Avalara Line Items
Add a liTaxIncluded field to the LineItem type in the Avalara module for
indicating that a line item's liTotalAmount field already includes the
tax amounts. This will be useful when tax transaction creation fails
during the Place Order process and we need to record tax after the
Customer has already been charged.

Refs #1525