[#1145] Add Google Analytics Tracking Script
Add our Google Analytics tracking script to the generated HTML file. Add
an outgoing port for logging new URLs & page titles.

Add a pageLoadCompleted function that determines whether the data for
the current page has been successfully loaded or not. Add an
updateWrapper that wraps the Main.update function, checking to see if
the page has finished loading in the latest update cycle. If so, use the
new port to log the page URL & title with Google Analytics.

Refs #1145