[#1145] Log Purchases to Google Analytics
When the checkout process is successfully completed, encode the returned
Order data into the Analytics' `purchase` event data. Send the data out
to a port so it can be logged with Google Analytics.

Add the `sku` field to the OrderProduct type. Expose the OrderProduct &
OrderLineItem types & decoders from the PageData module.

Closes #1145