[#1141] Add Dynamic Meta Attributes to Pages
Add Open Graph Protocol meta tags as well a canonical link tag to the
generate HTML file.

Add an optional image URL parameter to the logPageView port & rename it
to updatePageMetadata. Modify the javascript port code so that it not
only sends a page view to Google Analytics, but also updates the OGP
meta tags and the page's canonical URL.

Add a View.pageImage function for getting a Route's image. Currently
only the Product & Category routes have relevant images in their page
data.

Closes #1141