[#1603] Improve Performance of the Data Migration Script
Add an address cache to the DataMigration script, keeping Maps of
Addresses to their IDs(for preventing insertion of duplicate addresses)
and of Customer IDs to whether or not they have a default shipping or
billing address.

Modify the insertAddresses function so that it builds & returns this
cache. Modify the insertOrders function so it uses and builds the cache
as it inserts Orders into the new site's database.

Modify the insertUniqueAddress function of the DataMigration script so
that it queries the caches when checking for an existing or default
address, and updates the caches when inserting a new address.

These changes keep us from having to make over a million queries and
decreases the runtime of the migration script from over 24 hours to
about 30 minutes.

Closes #1603