[#1533] Ensure Migrated Data Retains It's ID Number
Modify the DataMigration script so that the Category, Product, Pages,
Customers, & Orders retain their old ID values. This simplifies the
insertions, reduces the number of ID Maps we needs to build, and lets us
redirect old URLs to the new URLs.

Note that the `insertKey` function from Persistent does not update the
ID sequence for the table, so after migrating the data  the script will
set the ID sequence value so the maximum ID for the table.

Closes #1533