[#1183] Add Redirecting API Routes For Migrating URLs
Add a Routes.Redirects module to the API server, containing routes for
Categories, Products, & Pages that take in an ID number and redirect the
user to the appropriate slug-based Client route.

This allows us to set our web server up to redirect pages from the old
website to the new website since the old website's URLs contained the
Category/Product/Page IDs in them.

Closes #1183