[#1615] Remove the IsActive Field From Products
Remove the isActive field from the Product's DB model. We only want to
allow de-activation of Variants to prevent confusion on the admin pages.
Instead, a Product's active status is derived from it's Variants active
status.

Rework various queries that check for active products so that they use
the Routes.Utils.activeVariantExists function to check if any of the
Product's Variants are active.

Remove the "Is Enabled" form field from the Create/Edit Product Admin
Page.

Closes #1615