[#1612] Add Fallback Image for Products & Catgories
Add a default image to show if a Product or Category has no defined
image.

Modify the ImageSourceSet type so that the original image is returned as
Nothing if the passed filename is blank. When decoding the ImageData in
the client, fallback to the default image if there is no original.

Closes #1612