[#1607] Add Tax Calculation for Virginia
Re-add the tax calculations, but just for Virginia and apply it to the
entire Order instead of just Products.

Tax calculations were originally removed since we had planned on using
Avalara for tax calculations. But we won't be able to use that on our
live site for at least a month as our application is reviewed & our
implementation is tested & verified. So this is a temporary patch to
make sure we still charge tax for our home state.

Closes #1607