[#1610] Move Additional Product Categories to Separate DB Table
Make a new ProductToCategory DB type that represents a many-to-many
joining table from products to categories. This replaces the list of
categories in the Product type, where we now only store the "main"
category for a Product. The main category is shown in the breadcrumbs &
category description for the Product's Details page.

Modify the paginatedSelect and related functions so they can be filtered
by the additional categories as well.

Change the DataMigration script to assign the master_categories_id as
the main category, and the remaining categories using the
ProductToCategory table.

Modify the Admin's Add/Edit Product routes so that multiple categories
can be set for a Product.

Use the main category for the Merchant Feed's category hierarchy.

Refs #1610