[#1623] Add Build Configuration for Selecting Stripe Key
Modify the client's webpack configuration so that setting the
`STRIPE_ENV` environmental variable to `production` will use the live
client key for Stripe. This leverages Webpack's DefinePlugin to replace
all instances of STRIPE_API_KEY in the source with the value selected by
switching on the environmental variable.

Remove the duplicate definition of the Google Analytics Measurement ID
by adding a define to the webpack configuration.

Closes #1623