[#1624] Include Telephone Numbers in Addresses
Add a phoneNumber field to the Address database type as well as the
AddressData type. Migrate this information during in the DataMigration
script.

Modify the client's Address Form & associated functions to include a
phoneNumber field and include it in the address data sent to the API
server.

Actually fix the pesky Google Analytics CSP error by adding additional
allowed script & img domains.

Closes #1624