Fix GraphicsMagick Mogrify Calls
Change the Images.resizeWithGraphicsMagick function to account for the
fact that GraphicsMagick wants to write the entire path passed to it
into the specified -output-directory.

For example, if passed a `folder/img.jpg` path, it will attempt to save
the resulting image at `<output-directory>/folder/img.jpg`. This
conflicts with what ImageMagick had done, which was to output to
`<output-directory>/img.jpg`.

To fix this, we employ relative paths for the output directory & change
the working directory to the folder containing the original image while
running GraphicsMagick.