[#1582] Never Set Order Status to OrderReceived in DataMigration Script
Modify the DataMigration script's getOrderStatus function so that it
never returns the OrderReceived status, even for order's where that's
applicable.

We do this because the only Order's with that status are very old and
the OrderReceived status is used to see if an Order should be exported
to StoneEdge. These two facts combined means that the first syncing with
StoneEdge will result in a handful of very old orders to be exported.

It's not problematic because we will be syncing the old website with
StoneEdge immediately before migrating to the new site.

Refs #1582