[#1543] Fix Scrolling to Anchor Links on StaticPages
Fix a bug causing anchor links to be ignored when included in links to a
StaticPage.

To fix this, we parse the hash fragment when parsing the PageDetails
route. Then, when page loading is complete & the page has a fragment, we
scroll to the element with a name attribute matching the fragment.

This requires us to only scroll to the top of the page when a
StaticPage's data has been loaded if there is no fragment in the URL.
Otherwise the page will bounce up and down.

We also modify the data fetching after authorization, re-fetching the
page's data if we are on a page requiring Authentication, or the Cart or
Checkout pages.

Add a homePage value to the Routing module, for easily constructing a
Route to the Home page.

Add a scrollToName function to the Ports module, wrapping the
scrollToSelector function for scrolling to elements with a specific
`name` attribute.

Modify the `scrollToSelector` javascript port code, making it also
scroll to the element if it is below the user's current viewport -
instead of only when the element is above the user's window position.

Closes #1543