[#1563] Add Metatag for Telling Prerender to Throw a Server Error
When the page loading has completed, determine whether any requests for
the page's data completed successfully or not. If unsuccessful, assign
the error a status code and use the new logStatusCode port to signal to
the Prerender server that it should throw an HTTP error when returning
the page contents.

Add a logStatusCode Port that sets the `prerender-status-code` meta
tag's content to the proper status code when fired.

Refactor the pageLoadCompleted function, pulling out the part that
operates generically on the current page's WebData, so that we can
re-use it for determining the page's status code.

Closes #1563