[#1627] Generate a robots.txt File for Staging & Production Environments
Use the `robotstxt-webpack-plugin` to generate a `robots.txt` file for
the client.

The plugin uses the STRIPE_ENV instead of the NODE_ENV usually used for
testing if we are on production, since we only want crawlers to hit
sites where we accept payments. In a production environment, it will
allow all URLs except for the `/admin/` site, while staging environments
will disallow all crawling.

Refs #1627