Prevent Adding SoldOut or Inactive Products to Cart
Fix a bug allowing Customers to add Inactive or Sold Out ProductVariants
to their Carts via the Quick Order page. This was caused by not
validating the current quantity or active status of the ProductVariants
being added to Carts.

Attempting to add an inactive or sold out variant will now return an
error message.