[#1604] Add Admin Page for Creating Coupons
Add a CouponNew AdminRoute, responsible for rendering, updating, &
submiting a form for Admins to create new Coupons with.

Add a JSON encoder for CouponTypes.

Refactor the selectRow function the HorizontalForm module, pulling out
the `select` element rendering into a separate `selectElement` function,
which eases re-use for forms with selects that share a row with another
input.

Refs #1604