[#1604] Add an Edit Coupon Page to the Admin Site
Add a new CouponEdit Admin Route, responsible for rendering, updating,
validating, & submitting a form for Admin user's to edit the details of
a Coupon.

Add an applyMaybe function to the Validation module for validating a
Maybe value only when an inner value exists.

Fix the CouponType validation for the CouponAdmin's NewForm.

Refactor the dateRow & couponTypeRow view helpers to make them
Msg-agnostic, allowing re-use in the edit view.

Closes #1604