[#1631] Add Script to Merge/Export Customers by Email
Add a MergeCustomers script that finds all Customers that have identical
emails on a case-insensitive basis. For common domains, it will merge
the Customers into a single Customer, along with their Addresses, Carts,
PasswordResets, Orders, & Reviews. For less common domains, it exports
the emails to a CSV file.

Rename the `mergeCarts` function in the Models.Utils module to
`mergeAnonmergeAnonymousCart`. Add a new `mergeCarts` function for
merging two generic carts together, instead of a specific Anonymous &
Customer cart. Tweak the type signature to allow running the function in
a script instead of limiting it to server route handlers.

Refs #1631