[#1631] Make the Customer & Checkouts Routes Case-Insensitive
Modify the Customer & Checkout routes so that they do not require
case-sensitivity in the Customer Email field. For validation, they
ensure that no matching email exists(ignoring the casing) & for querying
customers they do so case-insensitively.

Add a getCustomerByEmail function to the Models.Utils module for
fetching the first customer with a matching email in a case-insensitive
manner.

Add a noMatches function to the Valiation module for ensuring there is
no existing database entry matching the given filters.

Closes #1631