[#1132] Improve the Loading Screen Between Pages
Modify the "Loading..." interstitial text between page loads so that it
takes up the entire window, is translucent, fades in, and contains an
animated spinner.

Fix the lag between changing Category pages and the updating of the
Category details. This requires bumping the version of the paginate
library and conditionally resetting the Paginated type's response data.

Refs #1132