[#1642] Fix Rotation of Homepage's Carousel
Fix a bug preventing the Homepage's banner carousel from rotating.

We added a new initializeOrDestroyHomepageCarousel Port that is sent
a boolean indiciting if we are on the homepage or not. If on the
homepage, we initialize the Bootstrap Carousel and store a reference to
the element. When we are not on the homepage & we have a reference to
the element, we destroy the Carousel hooks on the stored element and
remove the reference.

Closes #1642