[#1641] Add BCCs & Sender Name to Emails
Add a BCC to the customer service email address when we send the Order
Confirmation email to customers.

Modify the email building & sending code so that we can add names to the
From, To, & BCC fields of emails we send. Currently only the OrderPlaced
email has a name attached to the recipient since we do not store a name
for customers. However we could add optional names to the other emails
by checking for a default address.

Closes #1642