[#1550] Add Lightboxes for Product Images
Modify the Product Details page so that a pop-up lightbox is shown when
a Customer clicks the product's image. The state of the lightbox is
stored in the global Model and is reset when receive the details data
for a Product.

This uses the `prikhi/bootstrap-gallery` package that was extracted from
the FIC website repository.

Closes #1550