[#1648] Add Link to Login Page on Anonymous Checkout Page
Rename the "Login Details" section of the Checkout page to "Create an
Account" when the page is displayed for a Customer that is not logged
in. Also include a link to the Login page next to the section heading.

This should clear up some ambiguity with initial fields of the form &
help out Customers who had confused this for a login form.

Closes #1648