Add a Generic Queue-Based Worker Pool Module
Add an ImmortalQueue module to the server which wraps the 'immortal'
package into a pool of unkillable threads for processing a queue of
actions.

The ImmortalQueue type defines the specifics of processing your queue:
the worker count, idle-worker polling time, how to pop/push an item off
the queue, how to process each item, and what to do when the processing
fails.

When processing of an ImmortalQueue starts, the worker threads are
branched off and a cycle of pop-the-queue & feed-the-workers is started.

There are exposed functions for closing a queue and killing it. Closing
a queue allows the job processing to finish, while killing it will cause
all jobs to exit immediately.

This is not hooked up to the server, it is simply the re-usable base for
implementing our own worker pool. With tests & more documentation, it
could be split off into a separate package.