[#1157] Add a Job Table to the Database
Add a new Job type for storing unprocessed asychronous tasks. A Job has
an action which is encoded as JSON, a queued time, a retry count, and an
optional time for it to be run.

Add a Models.PersistJSON module to define a type that allows us to
leverage PostgreSQL's ability to store JSON.

We store the tasks as JSON so that the database field can contain a type
that references the various ID fields of other tables. The alternative
approach would be to split the database & field definitions into two
sets so that the field for the Job table could reference other items in
the job table.

Refs #1157