[#1157] Add Worker Pool for Processing Enqueued Jobs
Add a new Workers module to the server. The module defines an
`ImmortalQueue` for processing jobs from the database, the `Task` type
that gets encoded into JSON and saved in the jobAction field, and a
helper function for enqueueing a Task.

Modify the Checkout, Customer, & Product Admin routes so that they
enqueue email & image tasks instead of performing the actions
themselves.

Modify the server's Main module so that it spins up a job queue
processor with 2 worker threads.

Refactor the Emails module, moving the EmailType values into an
EmailData type & re-purposing the EmailType for Tasks. Instead of
database values, the EmailType values now contain the identifiers to
pull the values out of the database. This allows the Task worker to
fetch the actual data using the getEmailData function. Modify the
`Emails.send` function so that it can be run in the IO monad.

Refactor the Emails.OrderPlaced module so that the fetchData function
can be run in an arbitrary SQL monad.

Refactor the Images module, splitting the scaleImage function into
functions for saving the original image, optimizing an image, &
scaling+optimizing an image. This allows route handlers to simply save
the original & allows the worker threads to do all optimization &
scaling.

Bump the SQL connection count to 2 in Development mode.

Refs #1157