Add Graceful Shutdown Handling to the Server
Modify the server so that it cleanly shuts down when sent a SIGTERM
signal.

This involves adding a signal handler that fills an MVar when triggered.
Then, instead of blocking by running the Warp server, we asynchronously
run the Warp server and block on reading the MVar. Once the MVar is
filled, we can stop the worker pool, web server, and properly close the
pools & log files.

This lets us ensure we never lose Tasks that are being processed when
restarting the server, since we can now call the killImmortalQueue
function.