[#1525] Refactor Refreshing of Checkout Details
Avalara requires us to submit at least a Street Address, Zipcode, &
State in order to get a sales tax quote. This commit changes the message
raising & details refreshing code to reduce our AvaTax API calls &
ensure we have the required address fields before making an API request.

We add a FieldMsg type to the Address module, allowing us to specify if
an input field should update the model in an onChange or onInput event.
This allows us to not make an Avalara API call on every keystroke in the
Street or ZipCode fields.

We also modify the refreshDetails function in the Checkout module to
check for changes to the street or zipCode fields and to ensure the
fields are not blank.

Refs #1525