Fix Server Exceptions Caused by Synchronous Email Sending
Add a sendWithRetries function to the Emails module that catches errors,
waits 0.5 seconds, & retries email sending up to 5 times. If all the
retries also fail, it returns instead of raising an error. This fixes a
bug where the sendMimeMail2 function call would fail during order
processing and return an error to the customer instead of showing the
Order Confirmatin page.