Fix Avalara Refund Encoding & Amounts
Modify the ToJSON instance for the RefundTransactionRequest so that the
refundLines array is only present if the refundType is Partial. This
should fix API errors we were receiving from Avalara about the
refundLines field.

Change the orderRefundRoute so that a Full refund is never submitted to
Avalara. We want to always use the Partial refund type because Full
ignores any previous refunds we have made which can cause the refund
amount to be larger than the actual order total.