Default State/Province Dropdowns to an Empty Value
Modify the Address form so that the State/Province select dropdowns are
defaulted to an empty value instead of the first item in the list.

Customers were confused as to why their state was set to Alabama and
some forgot to change the state since it was already set, this should
prevent those issues.

Refactor the checkout form's validation into a separate function that
validates that Customers have selected a Shipping and/or Billing State
and have properly confirmed their desired password.