Refactor Filtering Items on Admin Pages
Add a searchInput function to the Views.Admin module to abstract the
idea of clientside filtering of Admin Page items. This function takes
the model, an input message to raise, and a function that matches a
items against a single search term. It returns an HTML element
containing the search input and a function for filtering the list of
items.

Use the new Admin.searchInput function for rendering & filtering on the
Product Admin's List page.