[#1665] Add Search Keywords to Products
Extend the Products database model to store a `keywords` String. This
field can contain additional words that will be searched from the site's
frontend, but will not show up in the description.

Modify the Admin's Product form to render & process a Search Keywords
input field. Modify the add/edit Product API routes to decode an
optional `keywords` field in the Product Data, defaulting it to the
empty string if not present.

Closes #1665