Generalize The Category Selects on Products Admin Page
Rename the AdminProductCategory type in the client to
AdminCategorySelect and move the rendering function for the Product
Admin's category selections to the Views.Admin module. This will allow
us to easily re-use the rendering code for other admin pages that need
to allow selection of multiple categories, like the Surcharges page.

Rename the ProductCategory type in the server to AdminCategorySelect and
move them to the Routes.CommonData module so they can also be re-used
for various admin pages. The constructor and validation functions have
also been moved to the CommonData module.