Move Login Parameters to CommonData Module
Refactor the Routes.Customers module, moving the LoginParameters type &
the password validation logic for logins into the Routes.CommonData
module.

This will allow re-using the type & processing for a login route in the
Routes.Checkout module. We do checkout logins in a separate route
because the customer login route merges the anonymous & customer cart
while the checkout login should overwrite the customer cart.