[#1672] Add Checkout Login API Route
Add a distinct login route to the Checkout API Routes. This route
overwrites the customer cart instead of merging the anonymous & customer
carts like the Customer API's login route. It also accepts the
CustomerDetailsParameters and returns a new CheckoutDetailsData for the
logged in user.

This route will be used to present a login form directly on the Checkout
page so we do not have to redirect users away from the checkout page if
they need to log in.

Refs #1672