[#1672] Add Login Form to the Checkout Page
Replace the "Already have an Account?" link on the Checkout page
that redirects to the Login page with a button that swaps out the
registration form with a login form.

The login form includes a remember me checkbox & a link to the password
reset page. There is a "Log In" button that lets the customer log into
their account before placing their order so that they can use any of
their existing addresses during checkout. The customer can also delay
logging in by filling out the address information and clicking the Place
Order or Pay with Credit Card buttons. This will first attempt to log
the user in, then proceed with the normal behavior of the button click.

Closes #1672