[#1653] Add Timestamps to Server Logs
Replace the LoggerSets in the Config type with TimedFastLoggers,
allowing us to include a timestamp in the logging output. Add a
timedLogStr function for standardizing the format of the log lines.

Replace the usage of pushLogStrLn throughout the code with the
timedLogStr function and modify the Main module to initialize the
TimedFastLoggers instead of LoggerSets.

Closes #1653