Refactor Password Validation Error Handling
Modify the validatePassword function in the Routes.CommonData module to
run in the AppSQL monad and to throw exceptions instead of server
errors. Add a handlePasswordValidationError function that transforms the
exceptions into server/validation errors.

This will be useful for using the validatePassword function in places
where we do not want to immediately throw a server error and wish to do
some processing after the password validation fails.