[#1674] Try Logging In Customers If Checkout Regsitration Fails
When a customer enters an existing email into the registration form of
the Checkout page, try logging them in with the entered email &
password. If the credentials are valid, we proceed with placing the
order instead of prompting them to log in instead. If the credentials
are invalid, we return the usual authorization error.

Remove email uniqueness validation from the
AnonymousPlaceOrderParameters type so that we can try authenticating
customers in the place order route.

Add a new NoCustomer value to the PasswordValidationError type, thrown
by validatePassword when the customer with the given email does not
exist. This allows us to proceed with the normal customer registration
functionality in the anonymousPlaceOrderRoute.

Extract an overwriteCustomerCart function from the Checkout API's
loginRoute for re-use in the anonymousPlaceOrderRoute.

Closes #1674