[#1675] Allow iframes in HTML Content
Replace the usage of the `xss-santizie` package's `sanitize` function
with a custom implement in the Routes.Utils module that also allows
`iframe` elements. This is OK from a security perspective because we use
Content Security Policy directives in the client to specify the domains
allowed to be used as iframes.

Closes #1675