Add Validation Helpers for Fields Containing Lists
Add helper functions to the server's Validation module for validating
fields that contain lists of primitive types or complex sub-models.

The indexedValidation will apply a validation to a field containing a
list of items with multiple fields. The validation result will have all
of the sub-fields' names prefixed with an identifier and the index of
the item.

The indexedValidator function will apply a simple validation to a field
containing a list of single fields(e.g., a list of CategoryIds). The
validator passed to the function will return a list of validations for
that specific field and the function will turn this into a Validators
type by constructing the field name for each item from an identifier and
the item's index.