Add Array Validation Helpers to the Client
Add an `array` function to the client's Validation module for building a
FormValidation from a field containing an Array of a type with multiple
sub-fields. This will map over the array, adding a prefix and the index
to every error's field name.

Add an `applyValidation` function that is similar to the `apply`
function but applies a FormValidation(containing multiple sub-fields)
instead of a simple Validation(containing a single field).

Add some un-exported helper functions like map2FormValidation and
prefixFields to the Validation module.