[#1684] Add a Global Settings Table & Cache
Add a new Settings database table containing global configuration data,
as well a querying & manipulation functions that ensure only one row of
settings ever exists in the database. The table currently contains
settings fields for temporarily disabling the Checkout process.

Add a new getSettingsCache field to the Caches type so that settings can
be queried without a database call. Add writeSetting function to he
Server module for updating a setting, propagating the changes to both
the cache and database.

Refs #1684