[#1684] Use Settings to Toggle Checkout Processing
Re-enable the previously disabled checkout processing, but use the new
Settings table & disable checkout fields to determine whether order
placing is enabled and what message to render when it is disbled.

Note that we still need to add a Settings page to the admin to let
managers control the checkout status.

Refs #1684