[#1661] Add New Product Sale Page to Admin
Add an Admin page for creating new Product Sales, along with a button to
this page on the Product Sales List page.

Add a `price` field to the SaleProductData type.

Add Admin API routes for fetching the data necessary to render the New
Product Sale form as well as for submitting a new Product Sale.

Refs #1661