[#1661] Add Edit Product Sale Admin Page
Add a Edit Product Sale page to the Admin site. Add links to the page in
the Product Sales table and redirect to the Edit Page after successfully
creating a new Product Sale.

Move the posixToDateString function from the CouponAdmin module to
Models.Utils for reuse in the ProductSalesAdmin module.

Rename the AdminListProductSale client type & ListProductSales server
type to AdminProductSale & ProductSaleData to match with the fact that
they are used in both the List & Edit pages.

Switch the argument order of the toVariantData server function to
simplify creation of the SaleVariantData type.

Refs #1661