[#1121] Cache the Daily & Monthly Sales Reports
Modify the Cache type to store the data required for the Admin
Dashboard's daily & monthly sales reports. Move the report generation
code to from the Dashboard module to the Cache module, modifying the
route to simply fetch the data from the cache.

Add two new asynchronous tasks - one for recording new Orders in the
Sales Reports and one for adding a new day to the Sales Reports. Modify
the Checkout API's Place Order routes so they enqueue a task for
recording the order in the sales reports. The new day task will
re-enqueue itself to run at midnight local time time of the next day.

Closes #1121