[#1685] Add Ability to Disable Priority S&H
Add a new isPriorityEnabled field to the ShippingMethod model,
signifying if Priority S&H is currently available for the method.
Re-enable the priority shipping form on the Checkout page, but hide the
form if the method's priority S&H has been disabled.

Add a checkbox to the ShippingMethods admin page.

Fix a bug causing issues when de-activating shipping methods from the
admin page.

Closes #1685