[#1530] Add Shipping Restrictions to Products
Add a new shippingRestrictions field to Products, containing a list of
Regions that the product cannot be shipped to.

Add a muti-select to the Admin's Product pages for setting the shipping
restrictions to a US States, Territories, & Armed Forces addresses.

Show an error on the Checkout page when a Customer tries to place an
order to a restricted state.

Refs #1530