Fix Checkout Page for Customer's With Restricted Default Address
Fix a bug causing the Checkout page to not load properly when a
Customer's default Address conflicted with a Product's shipping
restrictions.

Instead of having a restriction error throw a validation error in the
details routes, we return the error with the details data. This allows
us to render the form with the initial data but also includes the
restrictions error for their default address.