Add Variant Parameter to Product Details Page, Fix Merchant URLs
Add a new `variant` query parameter to the Product Details URL. If
specified, the cart form will select the given variant when the page is
loaded.

This fixes an issue with Google Merchant, where they were parsing the
price of the first variant for all the variants, causing all additional
variants to be disapproved due to inaccurate prices.