Fix Taxing Exempt Customers
Fix a bug causing tax-exempt customers to see sales tax charges when
checking out. This was caused by always sending a customer code of
`SALES_ORDER` when asking Avalara for a tax quote. This was fixed by
accepting an optional CustomerCode & defaulting back to `SALES_ORDER` if
the quote is for an anonymous user or for authorized user with no
CustomerCode.