Redirect to Login Page When Checking Out
Modify the Checkout button on the Cart page so that it links to the
Login page with a redirect to the Checkout page when a Customer is not
logged in.

Add a new clearCart query flag to the Login Page & Login API Route. The
presence of this flag indicates that the server should clear the
Customer's cart before merging the Anonymous cart into the Customer's
account upon a successful login.