[#14] Add email verification after registration
Problem: We don't check if an email submitted by
a person is real or not. That prevent us from
making anonymous checkouts, because everyone may
create an account with an arbitrary email address

Solution: send a confirmation email when a user want
to register; do not allow them to log in before they
verify their email account. Suggest them to resend a
confirmation email when they attempt to log in.