[Chore] Use 'persistent-migration' DB migrations management
Project has got to a point when migrations generated by 'mkMigrate'
no longer coverall use-cases and we need to provide non-trivial
migrations to convert data from one schema to another.

Migrations are now done via 'persitent-migration'.
Currently, there is only one migration that sets the inital DB scheme.

Migrations are applied on the application start so that only missing
migrations are done on launch.

Application additionally checks that persistent scheme matches
the DB and exits with an error if it doesn't.