[#58] Support multiple pictures per product on the backend
We need to be able to store and show multiple pictures per product.

To support this, the DB was migrated to store multiple product image
URLs as 'PersistentList' and API was updated to handle the data for
multiple images.

Frontend was adjusted to peek the first returned image or show the
blankImage if the list is empty.