[#64] Support inventory policies for product variants
Each product variant now has one of the following inventory policies:
1) 'RequireStock' - the purchase is only allowed when the required
   amount of the product variant is available in stock.
2) 'AllowBackorder' - allow to purchase more than available in
   stock, but show a disclaimer regarding increased delivery delays
   to users when product variant QOH goes zero or below.
3) 'Unlimited' - allow to purchase regardless of QOH.

The disclaimer regarding limited availability and increased delivery
delays is shown in the product list, cart, and during checkout.

Implementation-wise, instead of relying on query parameters validation
(which returns HTTP code 422 with an error that is displayed to users),
"out of stock" error and "limited availability" warning are now returned
as a part of a 200 response to cart-related queries.