[#64] Validate stock for 'RequireStock' product variant on the backend
Problem: Currently, product inventory validation for products is only
provided on the frontend, which makes it possible to place an order with
"out of stock" item by doing plain API call outside of browser.

Solution: When creating an order, check if there are any out of stock
items with 'RequireStock' inventory policy and throw an error if there
are.