[#78] Make shipping address validation non-mandatory
In some cases, the address might be unverifiable, but USPS might still be
able to deliver packages there.

To support such addresses and reduce the number of errors in addresses
provided by customers, address verification error is shown to users, but
it's still possible to complete the action (edit an address, or proceed
with a checkout) even with unverified address.