[Chore] Disable address verificatio if 'POSTGRID_API_KEY' == ""
Docker compose still requires 'POSTGRID_API_KEY' env variable to mount a
secret.

We still want to be able to optionally disable address verifcation for
local development or staging.

In order to achieve this, the backend now checks if the given env
variable is an empty string and doesn't perform address verification if
this is the case.