Fetch worker jobs from the DB with timeout
Problem: In some cases workers get stuck. It's not entirely clear why
this happens, but one possible case is that 'qPop' gets stuck.

Solution: Run SQL queries in qPop with 5 seconds timeout. This should be
more than enough to find a job in the corresponding table, but at the
same time it should fail due to a timeout quickly enough if SQL query
got stuck.